The package was renamed Static methods moved to a new help class
